+++ include/swift/Runtime/Metadata.h
@@ -18,7 +18,7 @@
 #define SWIFT_RUNTIME_METADATA_H
 
 #include "swift/ABI/Metadata.h"
-#include "swift/Reflection/Records.h"
+#include "swift/RemoteInspection/Records.h"
 #include "swift/Runtime/Once.h"
 #include "swift/shims/Visibility.h"
