+++ lib/CMakeLists.txt
@@ -45,7 +45,7 @@ add_subdirectory(RemoteAST)
 add_subdirectory(Sema)
 add_subdirectory(Serialization)
 add_subdirectory(SwiftDemangle)
-add_subdirectory(SwiftReflection)
+add_subdirectory(SwiftRemoteInspection)
 add_subdirectory(SwiftRemoteMirror)
 add_subdirectory(SIL)
 add_subdirectory(SILGen)
