+++ stdlib/public/SwiftRemoteMirror/SwiftRemoteMirror.cpp
@@ -23,8 +23,8 @@ SWIFT_REMOTE_MIRROR_LINKAGE uint32_t swift_reflection_libraryVersion = 3;
 }
 
 #include "swift/Demangling/Demangler.h"
-#include "swift/Reflection/ReflectionContext.h"
-#include "swift/Reflection/TypeLowering.h"
+#include "swift/RemoteInspection/ReflectionContext.h"
+#include "swift/RemoteInspection/TypeLowering.h"
 #include "swift/Remote/CMemoryReader.h"
 #include "swift/Basic/Unreachable.h"
