+++ stdlib/tools/swift-reflection-test/CMakeLists.txt
@@ -3,7 +3,7 @@ add_swift_target_executable(swift-reflection-test BUILD_WITH_STDLIB
   overrides.c
   LINK_LIBRARIES
     swiftRemoteMirror
-    swiftReflection)
+    swiftRemoteInspection)
 # NOTE(compnerd) since _WINDLL has no impact on non-Windows targets,
 # we just spam it on all the targets for simplicity due to the build
 # structure of swift.  This will make the global variable imported,
