+++ tools/libSwiftScan/CMakeLists.txt
@@ -22,7 +22,7 @@ target_link_libraries(libSwiftScan PRIVATE
     swiftDependencyScan
     swiftDriverTool
     swiftStaticMirror
-    swiftReflection
+    swiftRemoteInspection
     swiftCompilerStub)
 
 set_target_properties(libSwiftScan
