+++ tools/swift-reflection-dump/CMakeLists.txt
@@ -6,4 +6,4 @@ add_swift_host_tool(swift-reflection-dump
 target_link_libraries(swift-reflection-dump
                       PRIVATE
                         swiftStaticMirror
-                        swiftReflection)
+                        swiftRemoteInspection)
