+++ tools/swift-reflection-dump/swift-reflection-dump.cpp
@@ -16,9 +16,9 @@
 #include "swift/ABI/MetadataValues.h"
 #include "swift/Basic/LLVMInitialize.h"
 #include "swift/Demangling/Demangle.h"
-#include "swift/Reflection/ReflectionContext.h"
-#include "swift/Reflection/TypeRef.h"
-#include "swift/Reflection/TypeRefBuilder.h"
+#include "swift/RemoteInspection/ReflectionContext.h"
+#include "swift/RemoteInspection/TypeRef.h"
+#include "swift/RemoteInspection/TypeRefBuilder.h"
 #include "swift/StaticMirror/ObjectFileContext.h"
 #include "llvm/ADT/StringSet.h"
 #include "llvm/Object/Archive.h"
