+++ unittests/Reflection/CMakeLists.txt
@@ -9,6 +9,6 @@ if(("${SWIFT_HOST_VARIANT_SDK}" STREQUAL "${SWIFT_PRIMARY_VARIANT_SDK}") AND
       SWIFT_INLINE_NAMESPACE=__runtime)
     target_link_libraries(SwiftReflectionTests
       PRIVATE
-      swiftReflection${SWIFT_PRIMARY_VARIANT_SUFFIX})
+      swiftRemoteInspection${SWIFT_PRIMARY_VARIANT_SUFFIX})
   endif()
 endif()
